refactor: move service setup to coordinator helper
- Created `custom_components/meraki_ha/core/coordinator_helpers/service_setup.py` and moved `async_setup_services` and schemas there.
- Added `async_setup_services` method to `MerakiDataUpdateCoordinator` in `coordinator.py` which delegates to the helper.
- Updated `__init__.py` to call `coordinator.async_setup_services`.
- Fixed `custom_components/meraki_ha/core/api/client.py` by removing dependency on deleted `DeviceFetcher` and integrating device fetching into `_async_fetch_initial_data`.

This completes the request to move service setup logic and also fixes a blocking issue in `client.py` where it was importing a non-existent module.
