+++ custom_components/meraki_ha/coordinator.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import logging
 from datetime import datetime, timedelta
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
@@ -25,6 +25,15 @@ from .core.api.client import MerakiAPIClient as ApiClient
 from .core.helpers import filter_ignored_networks, process_coordinator_data
 from .core.managers import AvailabilityTracker, PendingUpdateManager
 from .types import MerakiDevice, MerakiNetwork
+
+if TYPE_CHECKING:
+    from custom_components.meraki_ha.services.camera_service import CameraService
+    from custom_components.meraki_ha.services.device_control_service import (
+        DeviceControlService,
+    )
+    from custom_components.meraki_ha.services.switch_port_service import (
+        SwitchPortService,
+    )
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -317,3 +326,28 @@ class MerakiDataUpdateCoordinator(DataUpdateCoordinator[dict[str, Any]]):
 
         """
         self.availability_tracker.mark_check_done(network_id, "traffic")
+
+    async def async_setup_services(
+        self,
+        device_control_service: DeviceControlService,
+        switch_port_service: SwitchPortService,
+        camera_service: CameraService,
+    ) -> None:
+        """
+        Set up the services for the Meraki integration.
+
+        Args:
+        ----
+            device_control_service: The device control service.
+            switch_port_service: The switch port service.
+            camera_service: The camera service.
+
+        """
+        from .core.coordinator_helpers.service_setup import async_setup_services
+
+        await async_setup_services(
+            self.hass,
+            device_control_service,
+            switch_port_service,
+            camera_service,
+        )
